Move View/Widget/IdGeneratorTrait to View/Helper.
