+++ src/View/Helper/FormHelper.php
@@ -26,10 +26,10 @@ use Cake\View\Form\ContextInterface;
 use Cake\View\Form\EntityContext;
 use Cake\View\Form\NullContext;
 use Cake\View\Helper;
+use Cake\View\Helper\IdGeneratorTrait;
 use Cake\View\Helper\StringTemplateTrait;
 use Cake\View\StringTemplate;
 use Cake\View\View;
-use Cake\View\Widget\IdGeneratorTrait;
 use Cake\View\Widget\WidgetRegistry;
 use DateTime;
 use Traversable;
