+++ src/View/Helper/IdGeneratorTrait.php
@@ -12,7 +12,7 @@
  * @since         CakePHP(tm) v3.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\View\Widget;
+namespace Cake\View\Helper;
 
 use Cake\Utility\Inflector;
